Working on Geo Chat Flow & refactoring
Application:
- A new container class GeoChatContainer has been added;
- AppContainer has been expanded due to the preceding;

Fragments:
- GeoChatFragment has been updated;

Models:
- GeoChatUiState has been created;
- GeoChatViewModel has been expanded and refined too;

Domain:
- Common Message domain classes have been updated & restructured;
- A new class GeoChatUseCase has been created;

Data:
- Some unnecessary classes have been removed;
- A new repository class GeoMessageDataRepository has been created;

Other:
- Some other stuff has been provided as always;

+++ client/app/src/main/java/com/qubacy/geoqq/ui/screen/geochat/settings/GeoChatSettingsFragment.kt
@@ -230,7 +230,7 @@ class GeoChatSettingsFragment() : LocationFragment(), MapLoadedListener {
         if (locationPoint == null) return
 
         val locationCircle = Circle(
-            locationPoint, (mModel as GeoChatSettingsViewModel).getCurRadiusOptionMeters())
+            locationPoint, (mModel as GeoChatSettingsViewModel).getCurRadiusOptionMeters().toFloat())
 
         if (mCurLocationCircle != null) removeCircleFromMap(mCurLocationCircle!!)
